# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fantasy Premier League API",
        "description": "API documentation for Fantasy Premier League application",
        "contact": {
            "name": "API Support",
            "email": "support@fantasypremierleague.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.fantasypremierleague.com",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Registers a new user and returns an authentication token.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "description": "Authenticates the user and returns an API token.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Revokes the current access token of the authenticated user.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user",
                "description": "Returns the currently authenticated user's details.",
                "operationId": "7c1cf647d76fad4b3eb1549fa4fe2f4c",
                "responses": {
                    "200": {
                        "description": "User details returned successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all users (Admin only)",
                "description": "Returns a list of all users. Accessible only to admin users.",
                "operationId": "f7d2cad996dfe7336fc0fd10aef23475",
                "responses": {
                    "200": {
                        "description": "List of users returned successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get list of players",
                "description": "Returns a list of players with optional filters by name, position, or club. Supports pagination or returning all players.",
                "operationId": "a355f9879160729f824bdea53e739b05",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter players by first or second name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Haaland"
                        }
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "description": "Filter players by position (GKP, DEF, MID, FWD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "FWD"
                        }
                    },
                    {
                        "name": "club",
                        "in": "query",
                        "description": "Filter players by club name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Manchester City"
                        }
                    },
                    {
                        "name": "all",
                        "in": "query",
                        "description": "If set, returns all players without pagination",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of players per page (default 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of players",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Erling"
                                                    },
                                                    "second_name": {
                                                        "type": "string",
                                                        "example": "Haaland"
                                                    },
                                                    "position": {
                                                        "type": "string",
                                                        "example": "FWD"
                                                    },
                                                    "club_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "price": {
                                                        "type": "integer",
                                                        "example": 12000000
                                                    },
                                                    "total_points": {
                                                        "type": "integer",
                                                        "example": 50
                                                    },
                                                    "assists": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "goals": {
                                                        "type": "integer",
                                                        "example": 20
                                                    },
                                                    "expected_goals": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 18.5
                                                    },
                                                    "expected_assists": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 9.2
                                                    },
                                                    "gw1_points": {
                                                        "type": "integer",
                                                        "example": 6
                                                    },
                                                    "gw2_points": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "gw3_points": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "gw4_points": {
                                                        "type": "integer",
                                                        "example": 8
                                                    },
                                                    "gw5_points": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "club": {
                                                        "description": "Associated club",
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Manchester City"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No players found"
                    }
                }
            }
        },
        "/api/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get all teams for logged-in user",
                "description": "Returns a list of all teams belonging to the authenticated user",
                "operationId": "getTeams",
                "responses": {
                    "200": {
                        "description": "List of teams retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Dream Team"
                                            },
                                            "budget": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "players": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Create a new team",
                "description": "Create a new team for the authenticated user",
                "operationId": "createTeam",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Team name (must be unique)",
                                        "type": "string",
                                        "example": "My Dream Team"
                                    },
                                    "budget": {
                                        "description": "Team budget (optional, defaults to 100)",
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Team created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "My Dream Team"
                                        },
                                        "budget": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/teams/{team}/sync-players": {
            "put": {
                "tags": [
                    "Teams"
                ],
                "summary": "Sync team players",
                "description": "Replace all team players with a new list of 11 players",
                "operationId": "syncTeamPlayers",
                "parameters": [
                    {
                        "name": "team",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_ids"
                                ],
                                "properties": {
                                    "player_ids": {
                                        "description": "Exactly 11 player IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "maxItems": 11,
                                        "minItems": 11,
                                        "example": [
                                            1,
                                            2,
                                            3,
                                            4,
                                            5,
                                            6,
                                            7,
                                            8,
                                            9,
                                            10,
                                            11
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Team players synced successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Team updated successfully"
                                        },
                                        "team": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "budget": {
                                                    "type": "integer"
                                                },
                                                "players": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    },
                                                    "minItems": 11
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Team not found"
                    },
                    "422": {
                        "description": "Validation error - must provide exactly 11 players"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer"
            },
            "api_key": {
                "type": "apiKey",
                "description": "API Key authentication",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}